// src/notifications/enums/notification-type.enum.ts
export enum NotificationType {
  TASK_ASSIGNED = 'TASK_ASSIGNED',
  TASK_REASSIGNED = 'TASK_REASSIGNED',
  TASK_UPDATED = 'TASK_UPDATED',
  TASK_STATUS_CHANGED = 'TASK_STATUS_CHANGED',
  TASK_COMPLETED = 'TASK_COMPLETED',
  TASK_BLOCKED = 'TASK_BLOCKED',
  COMMENT_MENTION = 'COMMENT_MENTION',
  COMMENT_REPLY = 'COMMENT_REPLY',
  PROJECT_ASSIGNED = 'PROJECT_ASSIGNED',
  MILESTONE_ASSIGNED = 'MILESTONE_ASSIGNED',
  TEAM_MEMBER_ADDED = 'TEAM_MEMBER_ADDED',
  TEAM_MEMBER_REMOVED = 'TEAM_MEMBER_REMOVED',
  TEAM_MEMBER_TRANSFERRED = 'TEAM_MEMBER_TRANSFERRED',
  SYSTEM_NOTIFICATION = 'SYSTEM_NOTIFICATION',
}